refactored comment errors catching

+++ app/graphql/mutations/comments/delete_comment.rb
@@ -6,7 +6,10 @@ module Mutations
       field :message, String, null: false
 
       def resolve(comment_id:)
-        comment = current_user.comments.find(comment_id)
+        comment = Comment.find_by(id: comment_id)
+
+        raise GraphQL::ExecutionError, "Comment not found" unless comment.present?
+        raise GraphQL::ExecutionError, "You cannot delete this comment" unless comment.user_id == current_user.id
 
         comment.destroy
